+++ gulpfile.js
@@ -1,6 +1,7 @@
 const { series, parallel } = require("gulp");
 
 const images = require("./tasks/Image");
+const fonts = require("./tasks/Font");
 const styles = require("./tasks/Style");
 const scripts = require("./tasks/Script");
 const watch = require("./tasks/Server");
@@ -18,6 +19,14 @@ exports.imageResize = series(
     images.resizeLg3x,
     images.cachemin
 );
+
+exports.convertFonts = series(
+    fonts.transform,
+    fonts.transformToWoff2,
+    fonts.ttfRebase
+);
+
+exports.rebaseFonts = series(fonts.transform, fonts.ttfRebase);
 
 exports.convertImages = series(images.images, images.cachemin);